refactor: extract skill eval read core
