+++ packages/skill-evals/src/core/types.ts
@@ -1,0 +1,20 @@
+export type CommandResult = {
+  args: readonly string[];
+  command: string;
+  cwd: string;
+  exitCode: number;
+  stderr: string;
+  stdout: string;
+};
+
+export type RunPaths = {
+  binDir: string;
+  eventsPath: string;
+  packageRoot: string;
+  repoRoot: string;
+  runRoot: string;
+  shellEnvDir: string;
+  summaryJsonPath: string;
+  summaryMdPath: string;
+  workspacePath: string;
+};
