+++ packages/skill-evals/src/suites/first-tree-read/__tests__/metrics.test.ts
