+++ packages/skill-evals/src/suites/first-tree-read/cases.ts
