+++ packages/skill-evals/src/suites/first-tree-read/types.ts
@@ -1,3 +1,5 @@
+import type { CommandResult } from "../../core/types.js";
+
 export type WorkspaceKind = "blank" | "context-tree";
 
 export type FirstTreeReadEvalCase = {
@@ -8,27 +10,6 @@ export type FirstTreeReadEvalCase = {
   prompt: string;
   promptAlternates: readonly string[];
   workspaceKind: WorkspaceKind;
-};
-
-export type RunPaths = {
-  binDir: string;
-  eventsPath: string;
-  packageRoot: string;
-  repoRoot: string;
-  runRoot: string;
-  shellEnvDir: string;
-  summaryJsonPath: string;
-  summaryMdPath: string;
-  workspacePath: string;
-};
-
-export type CommandResult = {
-  args: readonly string[];
-  command: string;
-  cwd: string;
-  exitCode: number;
-  stderr: string;
-  stdout: string;
 };
 
 export type FixtureValidation = {
